Changed AdPreview to AdCreativePreview in AdCreative get_ad_preview method
Summary:
The AdCreative get_ad_preview method in the Python Ads Sdk currently iterates over AdPreview objects, whose endpoint is 'generatepreviews', instead of AdCreativePreview, whose endpoint is 'previews'

This is inconsistent with the PHP Ads Sdk, which iterates over AdPreview but with the 'previews' endpoint (the AdCreativePreview object does not exist in the PHP Ads SDK), and with https://developers.intern.facebook.com/docs/marketing-api/generatepreview/v2.3#examples

The first examples shows a call to the endpoint 'previews' which currently isn't supported by the Python Ads Sdk.

This is a blocking task for t6861580

Test Plan: create an AdCreative object and call get_ad_preview. You will currently get an error that the endpoint doesn't exist. Change AdPreview to AdCreativePreview and the method will now work.